# packages/server/openapi.yaml
openapi: 3.1.0
info:
  title: agentic-kanban REST API
  version: 0.1.9
  description: "Auto-generated from Hono route source via scripts/generate-openapi.ts. Schemas are inferred statically (best-effort) — request bodies come from the parseJsonBody<T> type argument; response bodies are untyped. Coverage is audited: see x-coverage for what is scanned and what is deliberately not described here."
servers:
  - url: /
    description: "Same-origin (default dev: http://localhost:3001)"
...
paths:
...
  /api/failure-patterns:
    get:
      summary: GET /api/failure-patterns — list all stored patterns
      tags:
        - failure-patterns
      operationId: get_api_failure-patterns
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
    post:
      summary: POST /api/failure-patterns — create a pattern manually
      tags:
        - failure-patterns
      operationId: post_api_failure-patterns
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema: {}
        "400":
          description: Error (thrown HTTPException, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
  /api/failure-patterns/ingest:
    post:
      summary: POST /api/failure-patterns/ingest — ingest a markdown file
      tags:
        - failure-patterns
      operationId: post_api_failure-patterns_ingest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
        "400":
          description: Error (thrown HTTPException, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
  /api/failure-patterns/search:
    get:
      summary: GET /api/failure-patterns/search?q=<text>&limit=<n> — find similar failures
      tags:
        - failure-patterns
      operationId: get_api_failure-patterns_search
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
  /api/failure-patterns/{id}:
    delete:
      summary: DELETE /api/failure-patterns/:id
      tags:
        - failure-patterns
      operationId: delete_api_failure-patterns_id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
  /api/flaky-tests:
    get:
      summary: GET /api/flaky-tests — list flaky tests
      tags:
        - flaky-tests
      operationId: get_api_flaky-tests
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: string
        - name: minRuns
          in: query
          required: false
          schema:
            type: string
        - name: windowDays
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
  /api/flaky-tests/parse:
    post:
      summary: POST /api/flaky-tests/parse — ingest test output JSON
      tags:
        - flaky-tests
      operationId: post_api_flaky-tests_parse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
        "400":
          description: Error (thrown HTTPException, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
  /api/flaky-tests/pin:
    delete:
      summary: DELETE /api/flaky-tests/pin — unpin a test
      tags:
        - flaky-tests
      operationId: delete_api_flaky-tests_pin
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
        "400":
          description: Error
          content:
            application/json:
              schema: {}
    post:
      summary: POST /api/flaky-tests/pin — pin a test as known-flaky
      tags:
        - flaky-tests
      operationId: post_api_flaky-tests_pin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
        "400":
          description: Error (thrown HTTPException, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
...
  /api/merge-queue:
    post:
      summary: |-
        * POST /api/merge-queue
           *
           * body: { workspaceIds: string[], dryRun?: boolean, skipOnConflict?: boolean }
           *
           * - dryRun: true  → returns JSON plan (sorted order + conflict matrix + per-workspace conflictPreviews)
           * - dryRun: false → streams SSE events while executing the queue
      tags:
        - merge-queue
      operationId: post_api_merge-queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
        "400":
          description: Error (thrown HTTPException, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
  /api/merge-queue/preview/{workspaceId}:
    post:
      summary: |-
        * POST /api/merge-queue/preview/:workspaceId
           *
           * Dry-run conflict preview for a single workspace. Read-only — does not mutate the worktree.
           * Returns: WorkspaceConflictPreview
      tags:
        - merge-queue
      operationId: post_api_merge-queue_preview_workspaceId
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
        "500":
          description: Error
          content:
            application/json:
              schema: {}
...
  /api/tags/merge:
    post:
      summary: POST /api/tags/merge — merge sourceIds into targetId, then delete sources
      tags:
        - tags
      operationId: post_api_tags_merge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
        "400":
          description: Error (thrown HTTPException, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
        "403":
          description: Error (thrown FORBIDDEN, status from domainErrorHandler)
          content:
            application/json:
              schema: {}
...
  /api/workspaces/{id}/merge:
    post:
      summary: report success for a merge that had not happened yet.
      tags:
        - workspace-actions
      operationId: post_api_workspaces_id_merge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: async
          in: query
          required: false
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {}
        "202":
          description: Success
          content:
            application/json:
              schema: {}
...