# packages/server/src/routes/failure-patterns.ts
import type { Database } from "../db/index.js";
import { createRouter } from "../middleware/create-router.js";
import { parseJsonBody } from "../middleware/parse-body.js";
import { failurePatternBody, failurePatternIngestBody } from "./failure-pattern-body-schemas.js";
import { queryInt } from "../middleware/query-params.js";
import {
  findSimilarFailures,
  listPatterns,
  createPattern,
  ingestLearningFile,
  deletePattern,
} from "../services/failure-pattern.service.js";

export function createFailurePatternsRoute(database: Database) {
  const router = createRouter();

  // GET /api/failure-patterns — list all stored patterns
  router.get("/", async (c) => {
    return c.json(await listPatterns(database));
  });

  // GET /api/failure-patterns/search?q=<text>&limit=<n> — find similar failures
  router.get("/search", async (c) => {
    const q = c.req.query("q") ?? "";
    const limit = queryInt(c, "limit", { def: 3, min: 1, max: 10 });
    if (!q.trim()) return c.json([]);
    const matches = await findSimilarFailures(q, limit, database);
    return c.json(matches.map(m => ({
      pattern: m.pattern,
      score: m.score,
      matchedKeywords: m.matchedKeywords,
    })));
  });

  // POST /api/failure-patterns — create a pattern manually
  router.post("/", async (c) => {
    const body = await parseJsonBody(c, failurePatternBody);
    const pattern = await createPattern({
      title: body.title,
      errorClass: body.errorClass ?? null,
      description: body.description ?? null,
      rootCause: body.rootCause ?? null,
      fix: body.fix ?? null,
      sourceType: body.sourceType ?? "manual",
      sourceRef: body.sourceRef ?? null,
    }, database);
    return c.json(pattern, 201);
  });

  // POST /api/failure-patterns/ingest — ingest a markdown file
  router.post("/ingest", async (c) => {
    const body = await parseJsonBody(c, failurePatternIngestBody);
    const ingested = await ingestLearningFile(body.filePath, database);
    return c.json({ ingested });
  });

  // DELETE /api/failure-patterns/:id
  router.delete("/:id", async (c) => {
    await deletePattern(c.req.param("id"), database);
    return c.json({ ok: true });
  });

  return router;
}