# packages/server/src/routes/merge-queue.ts
import { streamSSE } from "hono/streaming";
import { createRouter } from "../middleware/create-router.js";
import { parseJsonBody } from "../middleware/parse-body.js";
import { mergeQueueBody } from "./merge-queue-body-schemas.js";
import { createMergeQueueService } from "../services/merge-queue.service.js";
import type { Database } from "../db/index.js";
import type { BoardEventSink } from "../services/board-events.js";
import type { SessionLauncher } from "../services/session.manager.js";
import { errorMessage } from "@agentic-kanban/shared/lib/error-message";

export function createMergeQueueRoute(
  database: Database,
  getSessionManager: () => SessionLauncher,
  options?: { boardEvents?: BoardEventSink },
) {
  const router = createRouter();

  const queueService = createMergeQueueService({
    database,
    boardEvents: options?.boardEvents,
    getSessionManager,
  });

  /**
   * POST /api/merge-queue/preview/:workspaceId
   *
   * Dry-run conflict preview for a single workspace. Read-only — does not mutate the worktree.
   * Returns: WorkspaceConflictPreview
   */
  router.post("/preview/:workspaceId", async (c) => {
    const workspaceId = c.req.param("workspaceId");
    try {
      const plan = await queueService.computePlan([workspaceId]);
      const preview = plan.conflictPreviews[0] ?? { workspaceId, hasConflicts: false, conflictingFiles: [], isStale: false };
      return c.json({ ok: true, preview });
    } catch (err) {
      return c.json({ ok: false, error: errorMessage(err) }, 500);
    }
  });

  /**
   * POST /api/merge-queue
   *
   * body: { workspaceIds: string[], dryRun?: boolean, skipOnConflict?: boolean }
   *
   * - dryRun: true  → returns JSON plan (sorted order + conflict matrix + per-workspace conflictPreviews)
   * - dryRun: false → streams SSE events while executing the queue
   */
  router.post("/", async (c) => {
    const body = await parseJsonBody(c, mergeQueueBody);

    if (body.dryRun) {
      const plan = await queueService.computePlan(body.workspaceIds);
      return c.json({ ok: true, dryRun: true, plan });
    }

    // Execute the queue and stream SSE events
    return streamSSE(c, async (stream) => {
      try {
        for await (const event of queueService.executeQueue(body.workspaceIds, {
          skipOnConflict: body.skipOnConflict ?? false,
        })) {
          await stream.writeSSE({ data: JSON.stringify(event) });
          if (event.type === "done") break;
          if (stream.closed) break;
        }
      } catch (err) {
        try {
          await stream.writeSSE({
            data: JSON.stringify({
              type: "error",
              workspaceId: "",
              issueNumber: null,
              issueTitle: "",
              error: errorMessage(err),
            }),
          });
          await stream.writeSSE({
            data: JSON.stringify({ type: "done", merged: [], failed: [], skipped: [] }),
          });
        } catch {
          // stream already closed
        }
      }
    });
  });

  return router;
}